Revert "added time_type and notification_type to process relative times (#29)"
This reverts commit 4c432ebe0225d51b88116d478dd3f42e7787b416.

+++ migrations/init.sql
@@ -2,7 +2,7 @@ CREATE EXTENSION IF NOT EXISTS "uuid-ossp";
 
 CREATE TABLE IF NOT EXISTS "user" (
     "id" UUID PRIMARY KEY DEFAULT uuid_generate_v4(),
-    "telegram_id" BIGINT NOT NULL UNIQUE,
+    "telegram_id" INTEGER NOT NULL UNIQUE,
     "username" VARCHAR(255) UNIQUE,
     "name" VARCHAR(255),
     "timezone" INTEGER NOT NULL DEFAULT 5,
